+++ pydantic_ai_slim/pydantic_ai/models/gemini.py
@@ -48,6 +48,7 @@ from . import (
     download_item,
     get_user_agent,
 )
+from ._google_common import GeminiUsageMetaData, metadata_as_request_usage
 
 LatestGeminiModelNames = Literal[
     'gemini-2.0-flash',
@@ -803,7 +804,7 @@ class _GeminiResponse(TypedDict):
 
     candidates: list[_GeminiCandidates]
     # usageMetadata appears to be required by both APIs but is omitted when streaming responses until the last response
-    usage_metadata: NotRequired[Annotated[_GeminiUsageMetaData, pydantic.Field(alias='usageMetadata')]]
+    usage_metadata: NotRequired[Annotated[GeminiUsageMetaData, pydantic.Field(alias='usageMetadata')]]
     prompt_feedback: NotRequired[Annotated[_GeminiPromptFeedback, pydantic.Field(alias='promptFeedback')]]
     model_version: NotRequired[Annotated[str, pydantic.Field(alias='modelVersion')]]
     vendor_id: NotRequired[Annotated[str, pydantic.Field(alias='responseId')]]
@@ -823,82 +824,9 @@ class _GeminiCandidates(TypedDict):
     safety_ratings: NotRequired[Annotated[list[_GeminiSafetyRating], pydantic.Field(alias='safetyRatings')]]
 
 
-class _GeminiModalityTokenCount(TypedDict):
-    """See <https://ai.google.dev/api/generate-content#modalitytokencount>."""
-
-    modality: Annotated[
-        Literal['MODALITY_UNSPECIFIED', 'TEXT', 'IMAGE', 'VIDEO', 'AUDIO', 'DOCUMENT'], pydantic.Field(alias='modality')
-    ]
-    token_count: Annotated[int, pydantic.Field(alias='tokenCount', default=0)]
-
-
-class _GeminiUsageMetaData(TypedDict, total=False):
-    """See <https://ai.google.dev/api/generate-content#UsageMetadata>.
-
-    The docs suggest all fields are required, but some are actually not required, so we assume they are all optional.
-    """
-
-    prompt_token_count: Annotated[int, pydantic.Field(alias='promptTokenCount')]
-    candidates_token_count: NotRequired[Annotated[int, pydantic.Field(alias='candidatesTokenCount')]]
-    total_token_count: Annotated[int, pydantic.Field(alias='totalTokenCount')]
-    cached_content_token_count: NotRequired[Annotated[int, pydantic.Field(alias='cachedContentTokenCount')]]
-    thoughts_token_count: NotRequired[Annotated[int, pydantic.Field(alias='thoughtsTokenCount')]]
-    tool_use_prompt_token_count: NotRequired[Annotated[int, pydantic.Field(alias='toolUsePromptTokenCount')]]
-    prompt_tokens_details: NotRequired[
-        Annotated[list[_GeminiModalityTokenCount], pydantic.Field(alias='promptTokensDetails')]
-    ]
-    cache_tokens_details: NotRequired[
-        Annotated[list[_GeminiModalityTokenCount], pydantic.Field(alias='cacheTokensDetails')]
-    ]
-    candidates_tokens_details: NotRequired[
-        Annotated[list[_GeminiModalityTokenCount], pydantic.Field(alias='candidatesTokensDetails')]
-    ]
-    tool_use_prompt_tokens_details: NotRequired[
-        Annotated[list[_GeminiModalityTokenCount], pydantic.Field(alias='toolUsePromptTokensDetails')]
-    ]
-
-
 def _metadata_as_usage(response: _GeminiResponse) -> usage.RequestUsage:
     metadata = response.get('usage_metadata')
-    if metadata is None:
-        return usage.RequestUsage()  # pragma: no cover
-    details: dict[str, int] = {}
-    if cached_content_token_count := metadata.get('cached_content_token_count'):
-        details['cached_content_tokens'] = cached_content_token_count  # pragma: no cover
-
-    if thoughts_token_count := metadata.get('thoughts_token_count'):
-        details['thoughts_tokens'] = thoughts_token_count
-
-    if tool_use_prompt_token_count := metadata.get('tool_use_prompt_token_count'):
-        details['tool_use_prompt_tokens'] = tool_use_prompt_token_count  # pragma: no cover
-
-    input_audio_tokens = None
-    output_audio_tokens = None
-    cache_audio_read_tokens = None
-    for key, metadata_details in metadata.items():
-        if key.endswith('_details') and metadata_details:
-            metadata_details = cast(list[_GeminiModalityTokenCount], metadata_details)
-            suffix = key.removesuffix('_details')
-            for detail in metadata_details:
-                modality = detail['modality']
-                details[f'{modality.lower()}_{suffix}'] = value = detail.get('token_count', 0)
-                if value and modality == 'AUDIO':
-                    if key == 'prompt_tokens_details':
-                        input_audio_tokens = value
-                    elif key == 'candidates_tokens_details':
-                        output_audio_tokens = value
-                    elif key == 'cache_tokens_details':
-                        cache_audio_read_tokens = value
-
-    return usage.RequestUsage(
-        input_tokens=metadata.get('prompt_token_count', 0),
-        output_tokens=metadata.get('candidates_token_count', 0),
-        cache_read_tokens=cached_content_token_count,
-        input_audio_tokens=input_audio_tokens,
-        output_audio_tokens=output_audio_tokens,
-        cache_audio_read_tokens=cache_audio_read_tokens,
-        details=details,
-    )
+    return metadata_as_request_usage(metadata)
 
 
 class _GeminiSafetyRating(TypedDict):
